Header: Move `loginTask` into the `session` service
