+++ app/services/session.js
@@ -1,13 +1,15 @@
 import { alias } from '@ember/object/computed';
 import Service, { inject as service } from '@ember/service';
 
-import { task } from 'ember-concurrency';
+import { task, waitForEvent } from 'ember-concurrency';
+import window from 'ember-window-mock';
 
 import ajax from '../utils/ajax';
 import * as localStorage from '../utils/local-storage';
 
 export default class SessionService extends Service {
   @service store;
+  @service notifications;
   @service router;
 
   savedTransition = null;
@@ -26,6 +28,48 @@ export default class SessionService extends Service {
       localStorage.removeItem('isLoggedIn');
     }
   }
+
+  /**
+   * This task will open a popup window directed at the `github-login` route.
+   * After the window has opened it will wait for the window to send a message
+   * back and then evaluate whether the OAuth flow was successful.
+   *
+   * @see `github-authorize` route
+   */
+  @task(function* () {
+    let windowDimensions = [
+      'width=1000',
+      'height=450',
+      'toolbar=0',
+      'scrollbars=1',
+      'status=1',
+      'resizable=1',
+      'location=1',
+      'menuBar=0',
+    ].join(',');
+
+    let win = window.open('/github_login', 'Authorization', windowDimensions);
+    if (!win) {
+      return;
+    }
+
+    let event = yield waitForEvent(window, 'message');
+    if (event.origin !== window.location.origin || !event.data) {
+      return;
+    }
+
+    let { data } = event.data;
+    if (data && data.errors) {
+      this.notifications.error(`Failed to log in: ${data.errors[0].detail}`);
+      return;
+    } else if (!event.data.ok) {
+      this.notifications.error('Failed to log in');
+      return;
+    }
+
+    this.login();
+  })
+  loginTask;
 
   login() {
     this.isLoggedIn = true;
